speed-up, and refactoring all around
